test(taos): merge stmt perf tests
